Úprava testů a definice Find() a drobné dodělávky testů
  Úprava testů a definice Find()
    - Změna návratového datového typu z long na double.
    - Dodefinován rozsah vstupních hodnot na max 170!
    - Vytvořeny testy v souladu s novou definicí
  Dodefinováno chování v případě dělení nuly a nekonečna a naopak.

+++ src/CalculatorProject/MathLib/CalcMath.cs
@@ -48,7 +48,7 @@ namespace MathLib
     }
 
     // Vypočte faktorial z 'x'
-    public long Fact(Byte x)
+    public double Fact(Byte x)
     {
       return 0;
     }
@@ -59,7 +59,7 @@ namespace MathLib
       return 0;
     }
 
-    // Vypočte logaritmus z 'x' o základu 10 
+    // Vypočte logaritmus z 'x' o základu 10
     public double Log(double x)
     {
       return 0;
